# Callback chain refactoring and performance improvements