--- a/dash-renderer/src/APIController.react.js
+++ b/dash-renderer/src/APIController.react.js
@@ -94,7 +94,6 @@ const UnconnectedContainer = props => {
                         loadingMap
                     )}
                     _dashprivate_loadingStateHash={getLoadingHash(
-                        layout,
                         [],
                         loadingMap
                     )}