--- a/dash-renderer/src/AppProvider.react.tsx
+++ b/dash-renderer/src/AppProvider.react.tsx
@@ -2,25 +2,10 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import {Provider} from 'react-redux';
 
-import initializeStore, { observe } from './store';
+import initializeStore from './store';
 import AppContainer from './AppContainer.react';
 
-import executedCallbacks from './observers/executedCallbacks';
-import executingCallbacks from './observers/executingCallbacks';
-import isLoading from './observers/isLoading'
-import loadingMap from './observers/loadingMap';
-import prioritizeCallbacks from './observers/prioritizedCallbacks';
-import requestedCallbacks from './observers/requestedCallbacks';
-import storedCallbacks from './observers/storedCallbacks';
-
 const store = initializeStore();
-observe(isLoading);
-observe(loadingMap);
-observe(requestedCallbacks);
-observe(prioritizeCallbacks);
-observe(executingCallbacks);
-observe(executedCallbacks);
-observe(storedCallbacks);
 
 const AppProvider = ({hooks}: any) => {
     return (