@Marc-Andre-Rivet 2020-06-04
@alexcjohnson Addressed concerns about circular callbacks pruning vs. execution in https://github.com/plotly/dash/pull/1254/commits/c93389bf59f77aac3bcddeb8f4af13ccc34057e2 - with this commit circular callbacks will not simply be pruned out but will instead be executed with an arbitrary entrypoint callback and will trigger the callbacks in the loop until such a time that it would re-executed a previously executed callback (through the new `predecessors` prop)