@Marc-Andre-Rivet 2020-06-05
@alexcjohnson https://github.com/plotly/dash/pull/1254/commits/4396e7e8dc56db580cdb0927677d7a7c971d8b7e addresses a regression exposed by https://github.com/plotly/dash/commit/c93389bf59f77aac3bcddeb8f4af13ccc34057e2 - 

`ALLSMALLER` wildcard callbacks with lower ids were not blocking those high higher ids, each callback triggering new callbacks in higher ids, resulting in `n*(n+1)/2` callbacks instead of `n`. Also callbacks dependent on `ALL` on the same components as the `ALLSMALLER` callback above would be be executed too soon and result in incorrect results.