@alexcjohnson 2020-06-12
> @alexcjohnson [4396e7e](https://github.com/plotly/dash/commit/4396e7e8dc56db580cdb0927677d7a7c971d8b7e) addresses a regression exposed by [c93389b](https://github.com/plotly/dash/commit/c93389bf59f77aac3bcddeb8f4af13ccc34057e2) -
> 
> `ALLSMALLER` wildcard callbacks with lower ids were not blocking those high higher ids, each callback triggering new callbacks in higher ids, resulting in `n*(n+1)/2` callbacks instead of `n`. Also callbacks dependent on `ALL` on the same components as the `ALLSMALLER` callback above would be be executed too soon and result in incorrect results.

Ah, I see - that's the reason for the test failures in c93389b - but in principle I can imagine fixes for this issue that would still give the correct final state, just inefficiently. Would it be worth 🔒 that a little tighter by asserting the correct number of callback invocations in one or two of the existing [wildcard tests](https://github.com/plotly/dash/blob/dev/tests/integration/callbacks/test_wildcards.py)?

Beyond that, just my latest comment about `redux_state_rqs` https://github.com/plotly/dash/pull/1254#discussion_r439525925 and the reverted test comment above https://github.com/plotly/dash/pull/1254#discussion_r439527964 and I think we're ready to go!!!