@Marc-Andre-Rivet 2020-06-15
@alexcjohnson 
> Would it be worth 🔒 that a little tighter by asserting the correct number of callback invocations in one or two of the existing wildcard tests?

https://github.com/plotly/dash/pull/1254/commits/be36ab47ae5977e71a2a225175b353647e5f7531

Also, `redux_state_rqs` now supports both legacy and current implementation for Dash matrix testing with https://github.com/plotly/dash/pull/1254/commits/afcdba90b68c178303ec68140ae7a07005d9c4c3 and https://github.com/plotly/dash/pull/1254/commits/63f14afffe1c91d6a00b8a688da2a62f804a46e5